Fragmentize BluetoothActivity and fix Bluetooth connectivity logic [1/1]
[Problem]
The BluetoothActivity contained all the fragment classes as well as some of the Bluetooth connectivity logic cause the
class to become bloated.

On the LG G2 phone running 4.4.2, for the BluetoothAdapter to work properly, the adapter needs to be refreshed on every
Activity/Fragment onResume().

[Solution]
Move the fragments in the BluetoothActivity into their own classes and move the Bluetooth connectivity logic into the BluetoothCommunicator.

Create an open() and close() function in BluetoothCommunicator, which handles all the BluetoothAdapter idiosyncracies.